Add description and currency icon

+++ app/src/main/java/stetsenko/currencies/presenter/Rate.kt
@@ -1,3 +1,3 @@
 package stetsenko.currencies.presenter
 
-data class Rate(val abbreviation: String, val description: String, val value: String, val icon: Long)
+data class Rate(val abbreviation: String, val description: String, val value: String, val icon: Int)
